backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django backend for web-sale project"
requires-python = ">=3.11"
dependencies = [
    "django==5.2.5",
    "dependency-injector",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt[crypto]==5.3.0",
    "django-cors-headers==4.4.0",
    "drf-spectacular==0.27.2",
    "setuptools",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "coverage[toml]==7.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 80
target-version = "py311"

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]

[tool.ruff.lint]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DJ", "E501"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces, rather than tabs
indent-style = "space"

# Respect magic trailing commas for better formatting
skip-magic-trailing-comma = false

# Automatically detect the appropriate line ending
line-ending = "auto"

# Enable automatic line wrapping
docstring-code-format = true
docstring-code-line-length = 80

[tool.ruff.lint.isort]
known-third-party = ["django"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/virtualenv/*",
    "*/.venv/*",
    "*/tests/*",
    "*/test_*.py",
    "manage.py",
    "backend/settings.py",
    "backend/wsgi.py",
    "backend/asgi.py",
    "backend/urls.py",
    "backend/__init__.py",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "class.*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2
fail_under = 95.0

[tool.coverage.html]
directory = "htmlcov"